fix: bump scheduled tasks timeout to uint32 type

+++ src/domain/dto/createScheduledTask.go
@@ -6,7 +6,7 @@ type CreateScheduledTask struct {
 	Name        valueObject.ScheduledTaskName  `json:"name"`
 	Command     valueObject.UnixCommand        `json:"command"`
 	Tags        []valueObject.ScheduledTaskTag `json:"tags"`
-	TimeoutSecs *uint16                        `json:"timeoutSecs,omitempty"`
+	TimeoutSecs *uint32                        `json:"timeoutSecs,omitempty"`
 	RunAt       *valueObject.UnixTime          `json:"runAt,omitempty"`
 }
 
@@ -14,7 +14,7 @@ func NewCreateScheduledTask(
 	name valueObject.ScheduledTaskName,
 	command valueObject.UnixCommand,
 	tags []valueObject.ScheduledTaskTag,
-	timeoutSecs *uint16,
+	timeoutSecs *uint32,
 	runAt *valueObject.UnixTime,
 ) CreateScheduledTask {
 	return CreateScheduledTask{
